Created initial Customer model, controller, and repository
