+++ liftoff-project/src/main/java/org/launchcode/liftoffproject/models/AbstractEntity.java
@@ -11,8 +11,23 @@ public abstract class AbstractEntity {
     @GeneratedValue
     private int id;
 
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
     public int getId() {
         return id;
+    }
+
+    @Override
+    public String toString() {
+        return name;
     }
 
     @Override
