+++ liftoff-project/src/main/java/org/launchcode/liftoffproject/models/data/CustomerRepository.java
@@ -6,4 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+
 }
